Tests: Refactor handling of module .cpp tests
They are now centralized in `modules/SCsub` and compiled in `lib<name>_tests`
libraries.

So for most modules, adding new tests (whether .h or .cpp) just implies adding
files in the `tests` subfolder of their module, and the rest is automatic.

The only exception is GDScript as its test suite is a bit more intrusive and
requires some custom code in GDScriptLanguage and the LSP.
